+++ python/CHANGELOG.md
@@ -5,3 +5,24 @@
 | 1.5.4 | 2025-08-07 | Feat | add fronting domain option | tempookian | 28e62e5
 |  |  | Fix | client config was not updated after remote update | goingfine | 3f5b7bb
 |  |  | Refactor | update version handling and dependency | Tempookian | d4a03b4
+
+# CFSCanner (python) - Change Log
+
+| Version | Date | Type  | Message | Author | Commit |
+|---------|------|-------|---------|--------|--------|
+| 1.6.0a0 | 2026-01-30 | Feat | add fronting domain option | tempookian | 28e62e5
+|  |  | Fix | fix some logging issues | Tempookian | 6e66000
+|  |  |  | improve debug logs | Tempookian | 298fa15
+|  |  |  | client config was not updated after remote update | goingfine | 3f5b7bb
+|  |  | Refactor | update version handling and dependency | Tempookian | 55857ec
+
+# CFSCanner (python) - Change Log
+
+| Version | Date | Type  | Message | Author | Commit |
+|---------|------|-------|---------|--------|--------|
+| 1.6.0a1 | 2026-01-30 | Fix | fix log error | Tempookian | 9560674
+| 1.6.0a0 | 2026-01-30 | Feat | add fronting domain option | tempookian | 28e62e5
+|  |  | Fix | fix some logging issues | Tempookian | 6e66000
+|  |  |  | improve debug logs | Tempookian | 298fa15
+|  |  |  | client config was not updated after remote update | goingfine | 3f5b7bb
+|  |  | Refactor | update version handling and dependency | Tempookian | 55857ec